Refactoring some duplicate code
